requirements.txt
streamlit==1.41.1
python-dotenv==1.0.1

openai==1.58.1
langchain==0.3.14
langchain-openai==0.3.0

pypdf==5.1.0
rank-bm25==0.2.2
